Pushing minor changes, implementation of future signal, and basic buffer

+++ database/src/signal.rs
@@ -126,7 +126,7 @@ impl<T: Signal> SignalTracker<T> for SerialSignalTracker<T> {
  ****************************Signal*****************************
  ***************************************************************/
 
-pub type SignalId = u32; /* Type alias to describe signals */
+pub type SignalId = u64; /* Type alias to describe signals */
 const RECBUFFSIZE: usize = 512; /* Currently fixing size of record buffer */
